Fixes #344
