+++ src/auth/internal.ts
@@ -1,18 +1,20 @@
-import { impl } from "../setup/dependencies.js"
-import { InitOptions, State } from "../index.js"
+import * as authLobby from "./lobby.js"
+import { InitOptions } from "../init/types.js"
+import { State } from "./state.js"
+
 
 export const init = (options: InitOptions): Promise<State | null> => {
-  return impl.auth.init(options)
+  return authLobby.init(options)
 }
 
 export const register = (options: { username: string; email: string }): Promise<{success: boolean}> => {
-  return impl.auth.register(options)
+  return authLobby.register(options)
 }
 
 export const isUsernameValid = (username: string): Promise<boolean> => {
-  return impl.auth.isUsernameValid(username)
+  return authLobby.isUsernameValid(username)
 }
 
 export const isUsernameAvailable = (username: string): Promise<boolean> => {
-  return impl.auth.isUsernameAvailable(username)
+  return authLobby.isUsernameAvailable(username)
 }
