+++ src/index.ts
@@ -1,105 +1,15 @@
 import localforage from "localforage"
 
-import * as auth from "./auth/index.js"
+import * as auth from "./auth/internal.js"
 import * as common from "./common/index.js"
-import * as identifiers from "./common/identifiers.js"
-import * as pathing from "./path.js"
-import * as crypto from "./crypto/index.js"
 import * as ucan from "./ucan/internal.js"
-import * as ucanPermissions from "./ucan/permissions.js"
 
-import { Maybe } from "./common/index.js"
-import { Permissions } from "./ucan/permissions.js"
+import { InitOptions, InitialisationError } from "./init/types.js"
+import { State, scenarioContinuation, scenarioNotAuthorised, validateSecrets } from "./auth/state.js"
 import { loadFileSystem } from "./filesystem.js"
 
 import FileSystem from "./fs/index.js"
 
-
-// SCENARIO
-
-
-export enum Scenario {
-  NotAuthorised = "NOT_AUTHORISED",
-  AuthSucceeded = "AUTH_SUCCEEDED",
-  AuthCancelled = "AUTH_CANCELLED",
-  Continuation = "CONTINUATION"
-}
-
-
-
-// STATE
-
-
-export type State
-  = NotAuthorised
-  | AuthSucceeded
-  | AuthCancelled
-  | Continuation
-
-export type NotAuthorised = {
-  scenario: Scenario.NotAuthorised
-  permissions: Maybe<Permissions>
-
-  authenticated: false
-}
-
-export type AuthSucceeded = {
-  scenario: Scenario.AuthSucceeded
-  permissions: Maybe<Permissions>
-
-  authenticated: true
-  newUser: boolean
-  throughLobby: true
-  username: string
-
-  fs?: FileSystem
-}
-
-export type AuthCancelled = {
-  scenario: Scenario.AuthCancelled
-  permissions: Maybe<Permissions>
-
-  authenticated: false
-  cancellationReason: string
-  throughLobby: true
-}
-
-export type Continuation = {
-  scenario: Scenario.Continuation
-  permissions: Maybe<Permissions>
-
-  authenticated: true
-  newUser: false
-  throughLobby: false
-  username: string
-
-  fs?: FileSystem
-}
-
-
-
-// ERRORS
-
-
-/**
- * Initialisation error
- */
-export enum InitialisationError {
-  InsecureContext = "INSECURE_CONTEXT",
-  UnsupportedBrowser = "UNSUPPORTED_BROWSER"
-}
-
-
-
-// INTIALISE
-export type InitOptions = {
-  permissions?: Permissions
-
-  // Options
-  autoRemoveUrlParams?: boolean
-  loadFileSystem?: boolean
-  rootKey?: string
-}
 
 /**
  * Check if we're authenticated, process any lobby query-parameters present in the URL,
@@ -125,6 +35,7 @@ export async function initialise(options: InitOptions): Promise<State> {
   if (await isSupported() === false) throw InitialisationError.UnsupportedBrowser
 
   const state = await auth.init(options)
+
   // Allow auth implementation to return a State directly
   if (state) {
     return state
@@ -199,82 +110,3 @@ export * as ipfs from "./ipfs/index.js"
 export * as keystore from "./keystore.js"
 export * as machinery from "./common/index.js"
 export * as crypto from "./crypto/index.js"
-
-
-
-// ㊙️  ⚛  SCENARIOS
-
-
-export function scenarioAuthSucceeded(
-  permissions: Maybe<Permissions>,
-  newUser: boolean,
-  username: string,
-  fs: FileSystem | undefined
-): AuthSucceeded {
-  return {
-    scenario: Scenario.AuthSucceeded,
-    permissions,
-
-    authenticated: true,
-    throughLobby: true,
-    fs,
-    newUser,
-    username
-  }
-}
-
-export function scenarioAuthCancelled(
-  permissions: Maybe<Permissions>,
-  cancellationReason: string
-): AuthCancelled {
-  return {
-    scenario: Scenario.AuthCancelled,
-    permissions,
-
-    authenticated: false,
-    throughLobby: true,
-    cancellationReason
-  }
-}
-
-export function scenarioContinuation(
-  permissions: Maybe<Permissions>,
-  username: string,
-  fs: FileSystem | undefined
-): Continuation {
-  return {
-    scenario: Scenario.Continuation,
-    permissions,
-
-    authenticated: true,
-    newUser: false,
-    throughLobby: false,
-    fs,
-    username
-  }
-}
-
-export function scenarioNotAuthorised(
-  permissions: Maybe<Permissions>
-): NotAuthorised {
-  return {
-    scenario: Scenario.NotAuthorised,
-    permissions,
-
-    authenticated: false
-  }
-}
-
-
-export async function validateSecrets(permissions: Permissions): Promise<boolean> {
-  return ucanPermissions.paths(permissions).reduce(
-    (acc, path) => acc.then(async bool => {
-      if (bool === false) return bool
-      if (pathing.isBranch(pathing.Branch.Public, path)) return bool
-
-      const keyName = await identifiers.readKey({ path })
-      return await crypto.keystore.keyExists(keyName)
-    }),
-    Promise.resolve(true)
-  )
-}
