+++ src/init/types.ts
@@ -1,0 +1,22 @@
+import { Permissions } from "../ucan/permissions.js"
+
+
+/**
+ * Initialisation error
+ */
+export enum InitialisationError {
+  InsecureContext = "INSECURE_CONTEXT",
+  UnsupportedBrowser = "UNSUPPORTED_BROWSER"
+}
+
+
+
+// INTIALISE
+export type InitOptions = {
+  permissions?: Permissions
+
+  // Options
+  autoRemoveUrlParams?: boolean
+  loadFileSystem?: boolean
+  rootKey?: string
+}
