+++ src/setup.ts
@@ -1,5 +1,13 @@
 import { Endpoints, setup as internalSetup, UserMessages } from "./setup/internal.js"
 
+import { Implementation as AuthImplementation } from "./auth/implementation/types.js"
+import { Implementation as CryptoImplementation } from "./crypto/implementation/types.js"
+import { Implementation as StorageImplementation } from "./storage/implementation/types.js"
+
+import * as authImpl from "./auth/implementation.js"
+import * as cryptoImpl from "./crypto/implementation.js"
+import * as storageImpl from "./storage/implementation.js"
+
 
 /**
  * Toggle debug mode.
@@ -19,7 +27,6 @@ export function shouldPin({ enabled }: { enabled: boolean }): boolean {
   internalSetup.shouldPin = enabled
   return internalSetup.shouldPin
 }
-
 
 /**
  * Override endpoints.
@@ -42,8 +49,23 @@ export function endpoints(e: Partial<Endpoints>): Endpoints {
 }
 
 /**
+ * Set implementations.
+ *
+ * Override how auth, crypto or storage operates in webnative.
+ */
+export function setImplementations(opts: {
+  auth?: AuthImplementation
+  crypto?: CryptoImplementation
+  storage?: StorageImplementation
+}) {
+  if (opts.auth) authImpl.set(opts.auth)
+  if (opts.crypto) cryptoImpl.set(opts.crypto)
+  if (opts.storage) storageImpl.set(opts.storage)
+}
+
+/**
  * Configure messages that webnative sends to users.
- * 
+ *
  * `versionMismatch.newer` is shown when webnative detects
  *  that the user's filesystem is newer than what this version of webnative supports.
  * `versionMismatch.older` is shown when webnative detects that the user's
@@ -53,6 +75,3 @@ export function userMessages(m: Partial<UserMessages>): UserMessages {
   internalSetup.userMessages = { ...internalSetup.userMessages, ...m }
   return { ...internalSetup.userMessages }
 }
-
-
-export { setDependencies } from "./setup/dependencies.js"
